+++ darts/models/forecasting/arima.py
@@ -45,6 +45,7 @@ class ARIMA(TransferableFutureCovariatesLocalForecastingModel):
         trend: Literal["n", "c", "t", "ct"] | list[int] | None = None,
         random_state: int | None = None,
         add_encoders: dict | None = None,
+        min_train_length: int | None = None,
     ):
         """ARIMA
         ARIMA-type models extensible with exogenous variables (future covariates)
@@ -101,6 +102,10 @@ class ARIMA(TransferableFutureCovariatesLocalForecastingModel):
             ..
         random_state: int or None
             Controls the randomness for reproducible forecasting.
+        min_train_length
+            Optionally, set a custom minimum required training series length for this model to allow training on
+            shorter input series. By default, Darts sets a conservative minimum length to avoid downstream issues.
+            Changing this value might lead to such downstream issues. Default: ``None`` (keeps the default requirement).
 
         Examples
         --------
@@ -126,7 +131,7 @@ class ARIMA(TransferableFutureCovariatesLocalForecastingModel):
         ----------
         .. [1] https://numpy.org/doc/stable/reference/generated/numpy.poly1d.html
         """
-        super().__init__(add_encoders=add_encoders)
+        super().__init__(add_encoders=add_encoders, min_train_length=min_train_length)
         self.order = p, d, q
         self.seasonal_order = seasonal_order
         self.trend = trend
@@ -240,4 +245,4 @@ class ARIMA(TransferableFutureCovariatesLocalForecastingModel):
 
     @property
     def _target_window_lengths(self) -> tuple[int, int]:
-        return 30, 0
+        return self._min_train_input_length(30), 0