+++ darts/models/forecasting/sf_model.py
@@ -45,6 +45,7 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
         add_encoders: dict | None = None,
         quantiles: list[float] | None = None,
         random_state: int | None = None,
+        min_train_length: int | None = None,
     ):
         """StatsForecast Model.
 
@@ -124,6 +125,10 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
             with `num_samples > 1` or `predict_likelihood_parameters=True`.
         random_state
             Controls the randomness for reproducible forecasting.
+        min_train_length
+            Optionally, set a custom minimum required training series length for this model to allow training on
+            shorter input series. By default, Darts sets a conservative minimum length to avoid downstream issues.
+            Changing this value might lead to such downstream issues. Default: ``None`` (keeps the default requirement).
 
         Examples
         --------
@@ -173,7 +178,7 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
 
         # future covariates support can be added through the use of a linear model
         self._linreg: LinearRegressionModel | None = None
-        super().__init__(add_encoders=add_encoders)
+        super().__init__(add_encoders=add_encoders, min_train_length=min_train_length)
 
     def _fit(
         self,
@@ -392,7 +397,7 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
 
     @property
     def _target_window_lengths(self) -> tuple[int, int]:
-        return 10, 0
+        return self._min_train_input_length(10), 0
 
     @property
     def _supports_range_index(self) -> bool: